pyproject.toml
[project]
name = "eval-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "evalscope>=1.4.0",
    "flashinfer-cubin>=0.5.3",
    "flashinfer-jit-cache>=0.5.3",
    "flashinfer-python>=0.5.3",
    "hnswlib>=0.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]

[[tool.uv.index]]
url = "https://flashinfer.ai/whl/cu128"
